+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## UNRELEASED - tbc
+
+### Added
+
+- 2D positions extracted for military zone markers; not yet used in output 
+
+
 ## [1.0.0] - 2026-02-21
 
 ### Added